python: move Python detection to its own file
